Factor out common inference test methods
